move prefix prefill kernel test to a separate file
